chore(ml): Refactor ml error logic
